Switch target type, condition type and parameter type values to integers

+++ lib/Layout/Resolver/TargetType/Taxon.php
@@ -27,7 +27,7 @@ final class Taxon extends TargetType implements ValueObjectProviderInterface
     {
         return [
             new Constraints\NotBlank(),
-            new Constraints\Type(type: 'numeric'),
+            new Constraints\Type(type: 'int'),
             new Constraints\Positive(),
             new SyliusConstraints\Taxon(),
         ];
